configure delete API endpoint. HTTP method experimentation on one endpoint through Next JS resulted in failure.

+++ my-app/app/deleteAPI/[quizId]/page.tsx
@@ -11,15 +11,16 @@ export default async function DeleteAPI({params}) {
   if (!user) {
     return redirect("/login");
   }
-  const response = await supabase.from('quizzes').delete().eq('id', `${quizId}`).eq('user_id', `${user.id}`)
-  
+  if (quizId>=4) {
+      const response = await supabase.from('quizzes').delete().eq('id', `${quizId}`).eq('user_id', `${user.id}`)
+  }  
 
   return (
     <div className="flex-1 w-full flex flex-col gap-20 items-center">
       <div className="w-full">
         <div className="py-6 font-bold bg-purple-950 text-center">
           This is a protected page that you can only see as an authenticated
-          user
+          user.
         </div>
         <nav className="w-full flex justify-center border-b border-b-foreground/10 h-16">
           <div className="w-full max-w-4xl flex justify-between items-center p-3 text-sm">
